refactor: move errors to dedicated types file
