+++ src/index.ts
@@ -1,4 +1,13 @@
 export { AccountIdentifier, SubAccount } from "./account_identifier";
+export * from "./errors/governance.errors";
+export {
+  CouldNotClaimNeuronError,
+  GovernanceError,
+  InsufficientAmountError,
+  StakeNeuronError,
+  StakeNeuronTransferError,
+  UnsupportedValueError,
+} from "./errors/governance.errors";
 export { GovernanceCanister } from "./governance";
 export { ICP } from "./icp";
 export { LedgerCanister } from "./ledger";
