+++ src/types/governance_converters.ts
@@ -496,12 +496,3 @@ export interface MakeExecuteNnsFunctionProposalRequest {
   nnsFunction: number;
   payload: ArrayBuffer;
 }
-
-/**
- * An error used to ensure at compile-time that it's never reached.
- */
-export class UnsupportedValueError extends Error {
-  constructor(value: never) {
-    super("Unsupported value: " + value);
-  }
-}
